Cleanup after merge from master Relocate tests after refactored code and fix method visibility
